Add support for JetBrains IDEs (#148)
This commit adds a new jetbrains directory with the included
Dracula.icls file. This file was generated from the existing
Dracula.tmTheme using the JetBrains colorSchemeTool.

In theory, this color scheme should work for all JetBrains/IntelliJ
IDEs, but I have only been able to test it with PyCharm, where it
appears to be working nicely. We may want to find some other folks to
test it out in other JetBrains IDEs.